% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{data_summary}
\alias{data_summary}
\title{Summary of dataset}
\usage{
data_summary(data)
}
\arguments{
\item{data}{is the dataset that you want to digest. dataset class must be dataframe or tibble.}
}
\value{
Returns a summary of the dataset of each
}
\description{
Summary of dataset
}
\examples{
data_summary(mtcars)
}
