% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{dsetinfo}
\alias{dsetinfo}
\title{Information about datasets}
\usage{
dsetinfo(data)
}
\arguments{
\item{data}{is the dataset that you want to digest. dataset class must be dataframe or tibble.}
}
\value{
Returns information on variables, observation, missing values and uniquness.
}
\description{
Information about datasets
}
\examples{
destinfo(mtcars)
}
