% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{full_profile}
\alias{full_profile}
\title{Generates full profile of the dataframe passed}
\usage{
full_profile(data)
}
\arguments{
\item{data}{is the dataset that you want to digest. dataset class must be dataframe or tibble.}
}
\value{
Returns a full digestion(profile) of the data
}
\description{
Generates full profile of the dataframe passed
}
\examples{
full_profile(mtcars)
}
