% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{var_types}
\alias{var_types}
\title{Information on the variables}
\usage{
var_types(data)
}
\arguments{
\item{data}{is the dataset that you want to digest. dataset class must be dataframe or tibble.}
}
\value{
Returns the count of data_types.
}
\description{
Information on the variables
}
\examples{
var_types(mtcars)
}
