% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{vis_mdata}
\alias{vis_mdata}
\title{Visualization of missing data}
\usage{
vis_mdata(data)
}
\arguments{
\item{data}{is the dataset that you want to digest. dataset class must be dataframe or tibble.}
}
\value{
Return a graph that sows the variables and rows with missing data
}
\description{
Visualization of missing data
}
\examples{
vis_mdata(mtcars)
}
